+++ lib/packages/core/src/api/composeApi.ts
@@ -9,18 +9,26 @@ import {ApiVersion} from '../constants/apiVersion';
 import {ApiComposer} from './apiComposer';
 import {getBlockByHeight, getBlockById, getBlockByTimestamp, getBlockId, getBlocks} from './factories/block';
 import {getBlockchainStatus, getPeer, getPeers, getServerStatus, getTime, suggestFee} from './factories/network';
-import {sendTextMessage} from './factories/message';
-import {sendEncryptedTextMessage} from './factories/message';
+import {sendEncryptedTextMessage, sendTextMessage} from './factories/message';
 import {
-    generateSendTransactionQRCode, generateSendTransactionQRCodeAddress, getAccount,
-    getAccountBalance, getAccountBlockIds, getAccountBlocks,
-    getAccountTransactions, getAliases,
-    getUnconfirmedAccountTransactions, setAccountInfo, setAlias, setRewardRecipient
+    generateSendTransactionQRCode,
+    generateSendTransactionQRCodeAddress,
+    getAccount,
+    getAccountBalance,
+    getAccountBlockIds,
+    getAccountBlocks,
+    getAccountTransactions,
+    getAliases,
+    getUnconfirmedAccountTransactions,
+    setAccountInfo,
+    setAlias,
+    setRewardRecipient
 } from './factories/account';
 import {getAliasById, getAliasByName} from './factories/alias';
-import {getContractsByAccount, getContract, getAllContractIds} from './factories/contract';
+import {getAllContractIds, getContract, getContractsByAccount} from './factories/contract';
 import {
-    broadcastTransaction, getTransaction,
+    broadcastTransaction,
+    getTransaction,
     sendAmount,
     sendAmountToMultipleRecipients,
     sendAmountToSingleRecipient,
@@ -28,16 +36,27 @@ import {
 } from './factories/transaction';
 import {getAllAssets, getAsset} from './factories/asset';
 import {DefaultApiEndpoint} from '../constants';
+import {AxiosRequestConfig} from 'axios';
 
 
 /**
  * Settings for API used in [[composeApi]]
  */
 export class ApiSettings {
+    /**
+     * @param nodeHost {string} The url of the Burst peer
+     * @param apiRootUrl {string} The optional api endpoint, i.e. `/burst`. Usually, you wont use this.
+     * @param apiVersion {ApiVersion} For future usage.
+     * @param httpOptions {any | AxiosRequestSettings}   Optional http options, like additional header.
+     * The default implementation uses axios. In case of a custom client pass your own options.
+     * see [Axios Configuration](https://github.com/axios/axios#request-config)
+
+     */
     constructor(
         public nodeHost: string,
         public apiRootUrl: string = DefaultApiEndpoint,
-        public apiVersion: ApiVersion = ApiVersion.V1
+        public apiVersion: ApiVersion = ApiVersion.V1,
+        public httpOptions?: any | AxiosRequestConfig,
     ) {
     }
 }
@@ -49,7 +68,6 @@ export class ApiSettings {
  * ```ts
  * const api = composeApi({
  *   nodeHost: 'https://wallet1.burst-team.us:2083', // one of the mainnet nodes
- *   apiRootUrl: '/burst' // endpoint to the BURST API
  * })
  * ```
  *