+++ lib/packages/http/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+# 0.4.1
+- Adding configuration/options for Http clients
+
+
 # 0.4.0
 - Changed License: From GPL-3.0 to Apache 2.0